requirements.txt
python-dotenv
requests
flask
google-genai>=0.3.0